feat(forms): implement filters in summary
