+++ frontend/src/app/[locale]/events/[eventSlug]/surveys/[surveySlug]/summary/page.tsx
@@ -1,18 +1,15 @@
 import { notFound } from "next/navigation";
+import FieldSummaryComponent from "./FieldSummaryComponent";
 import { ReturnLink } from "./ReturnLink";
 import { gql } from "@/__generated__";
 import { getClient } from "@/apolloClient";
 import { auth } from "@/auth";
-import Linebreaks from "@/components/helpers/Linebreaks";
+import { DimensionFilters } from "@/components/dimensions/DimensionFilters";
+import { buildDimensionFilters } from "@/components/dimensions/helpers";
 import {
-  Field,
   Layout,
   validateFields,
   validateSummary,
-  FieldSummary,
-  Choice,
-  SelectFieldSummary,
-  TextFieldSummary,
 } from "@/components/SchemaForm/models";
 import SchemaFormField from "@/components/SchemaForm/SchemaFormField";
 import SignInRequired from "@/components/SignInRequired";
@@ -20,10 +17,14 @@ import ViewContainer from "@/components/ViewContainer";
 import ViewHeading from "@/components/ViewHeading";
 import getPageTitle from "@/helpers/getPageTitle";
 import { getTranslations } from "@/translations";
-import { Translations } from "@/translations/en";
 
 const query = gql(`
-  query SurveySummary($eventSlug: String!, $surveySlug: String!, $locale: String) {
+  query SurveySummary(
+    $eventSlug: String!,
+    $surveySlug: String!,
+    $locale: String,
+    $filters: [DimensionFilterInput!],
+  ) {
     event(slug: $eventSlug) {
       name
 
@@ -31,250 +32,22 @@ const query = gql(`
         survey(slug: $surveySlug) {
           title(lang: $locale)
           fields(lang: $locale)
-          summary
+          summary(filters: $filters)
+          countFilteredResponses: countResponses(filters: $filters)
           countResponses
+          dimensions {
+            slug
+            title(lang: $locale)
+            values {
+              slug
+              title(lang: $locale)
+            }
+          }
         }
       }
     }
   }
 `);
-
-interface TextFieldSummaryComponentProps {
-  fieldSummary: TextFieldSummary;
-  translations: Translations;
-}
-
-function TextFieldSummaryComponent({
-  fieldSummary,
-  translations,
-}: TextFieldSummaryComponentProps) {
-  const t = translations.Survey;
-  const { summary, countResponses, countMissingResponses } = fieldSummary;
-  return (
-    <>
-      {fieldSummary.summary.map((item, idx) => (
-        <div key={idx} className="card mb-2">
-          <div className="card-body p-2 ps-3 pe-3">
-            <Linebreaks text={item} />
-          </div>
-        </div>
-      ))}
-      <p className="text-muted">
-        {t.attributes.countResponses}: {countResponses}.{" "}
-        {t.attributes.countMissingResponses}: {countMissingResponses}.
-      </p>
-    </>
-  );
-}
-
-interface SelectFieldSummaryComponentProps {
-  translations: Translations;
-  choices: Choice[];
-  fieldSummary: SelectFieldSummary;
-  showMissingResponses?: boolean;
-}
-
-function SelectFieldSummaryComponent({
-  translations,
-  choices,
-  fieldSummary,
-  showMissingResponses,
-}: SelectFieldSummaryComponentProps) {
-  const { countResponses, countMissingResponses, summary } = fieldSummary;
-  showMissingResponses ??= true;
-  const t = translations.Survey;
-
-  return (
-    <table className="table table-striped table-bordered">
-      <thead>
-        <tr>
-          <th scope="col">{t.attributes.choice}</th>
-          <th scope="col">{t.attributes.percentageOfResponses}</th>
-          <th scope="col">{t.attributes.countResponses}</th>
-        </tr>
-      </thead>
-      <tbody>
-        {Object.entries(summary).map(([choiceSlug, countThisChoice], idx) => {
-          const choice = choices.find((c) => c.slug === choiceSlug);
-          const choiceTitle = choice?.title || (
-            <>
-              <span title={t.warnings.choiceNotFound}>⚠️</span>{" "}
-              <em>{choiceSlug}</em>
-            </>
-          );
-          const percentage = countResponses
-            ? Math.round((countThisChoice / countResponses) * 100)
-            : 0;
-
-          return (
-            <tr key={choiceSlug}>
-              <td className="align-middle" scope="row">
-                {choiceTitle}
-              </td>
-              <td className="align-middle" style={{ width: "30%" }}>
-                <div className="progress">
-                  <div
-                    className="progress-bar"
-                    role="progressbar"
-                    style={{
-                      width: `${percentage}%`,
-                    }}
-                    aria-valuenow={percentage}
-                    aria-valuemin={0}
-                    aria-valuemax={100}
-                  >
-                    {percentage}%
-                  </div>
-                </div>
-              </td>
-              <td className="align-middle" style={{ width: "10%" }}>
-                {countThisChoice}
-              </td>
-            </tr>
-          );
-        })}
-        <tr>
-          <td>
-            ❓{" "}
-            <em className="text-muted">{t.attributes.countMissingResponses}</em>
-          </td>
-          <td></td>
-          <td>
-            <span className="text-muted">{countMissingResponses}</span>
-          </td>
-        </tr>
-      </tbody>
-    </table>
-  );
-}
-
-interface FieldSummaryComponentProps {
-  translations: Translations;
-  field: Field;
-  fieldSummary: FieldSummary;
-}
-
-// NOTE: if you rename this to FieldSummary, it'll clash with models and graphql-typegen will silently fail
-function FieldSummaryComponent({
-  translations,
-  field,
-  fieldSummary,
-}: FieldSummaryComponentProps) {
-  const t = translations.Survey;
-
-  let choices: Choice[] = [];
-  let questions: Choice[] = [];
-
-  const { countResponses, countMissingResponses } = fieldSummary;
-
-  switch (field.type) {
-    case "SingleLineText":
-      if (field.htmlType === "number" && fieldSummary.type == "Select") {
-        choices = Object.keys(fieldSummary.summary).map((key) => ({
-          slug: key,
-          title: key,
-        }));
-      }
-      break;
-
-    case "SingleCheckbox":
-      choices = [
-        {
-          slug: "checked",
-          title: t.checkbox.checked,
-        },
-        {
-          slug: "unchecked",
-          title: t.checkbox.unchecked,
-        },
-      ];
-      break;
-
-    case "SingleSelect":
-    case "MultiSelect":
-      choices = field.choices;
-      break;
-
-    case "RadioMatrix":
-      choices = field.choices;
-      questions = field.questions;
-      break;
-  }
-
-  switch (fieldSummary.type) {
-    // TODO handle htmlType="number"
-    case "Text":
-      // TODO fix padding
-      return (
-        <TextFieldSummaryComponent
-          fieldSummary={fieldSummary}
-          translations={translations}
-        />
-      );
-
-    case "SingleCheckbox":
-      const singleCheckboxFieldSummary: SelectFieldSummary = {
-        ...fieldSummary,
-        type: "Select",
-        summary: {
-          checked: fieldSummary.countResponses,
-          unchecked: fieldSummary.countMissingResponses,
-        },
-      };
-
-      return (
-        <SelectFieldSummaryComponent
-          translations={translations}
-          choices={choices}
-          fieldSummary={singleCheckboxFieldSummary}
-          showMissingResponses={false}
-        />
-      );
-
-    case "Select":
-      return (
-        <SelectFieldSummaryComponent
-          translations={translations}
-          choices={choices}
-          fieldSummary={fieldSummary}
-        />
-      );
-
-    case "Matrix":
-      const countTotalResponses = countResponses + countMissingResponses;
-      return questions.map((question) => {
-        const questionSummary = fieldSummary.summary[question.slug];
-        const countQuestionResponses = Object.values(questionSummary).reduce(
-          (a, b) => a + b,
-          0,
-        );
-        const countQuestionMissingResponses =
-          countTotalResponses - countQuestionResponses;
-
-        const matrixFieldSummary: SelectFieldSummary = {
-          countResponses: countQuestionResponses,
-          countMissingResponses: countQuestionMissingResponses,
-          type: "Select",
-          summary: fieldSummary.summary[question.slug],
-        };
-
-        return (
-          <>
-            <div className="mb-1">{question.title}</div>
-            <SelectFieldSummaryComponent
-              key={question.slug}
-              translations={translations}
-              choices={choices}
-              fieldSummary={matrixFieldSummary}
-            />
-          </>
-        );
-      });
-
-    default:
-      return undefined;
-  }
-}
 
 interface Props {
   params: {
@@ -282,9 +55,12 @@ interface Props {
     eventSlug: string;
     surveySlug: string;
   };
+  searchParams: {
+    [key: string]: string;
+  };
 }
 
-export async function generateMetadata({ params }: Props) {
+export async function generateMetadata({ params, searchParams }: Props) {
   const { locale, eventSlug, surveySlug } = params;
   const translations = getTranslations(locale);
 
@@ -296,9 +72,11 @@ export async function generateMetadata({ params }: Props) {
 
   const t = translations.Survey;
 
+  const { from: _, ...filterSearchParams } = searchParams;
+  const filters = buildDimensionFilters(filterSearchParams);
   const { data } = await getClient().query({
     query,
-    variables: { locale, eventSlug, surveySlug },
+    variables: { eventSlug, surveySlug, locale, filters },
   });
 
   if (!data.event?.forms?.survey) {
@@ -315,7 +93,7 @@ export async function generateMetadata({ params }: Props) {
   return { title };
 }
 
-export default async function SummaryPage({ params }: Props) {
+export default async function SummaryPage({ params, searchParams }: Props) {
   const { locale, eventSlug, surveySlug } = params;
   const translations = getTranslations(locale);
   const t = translations.Survey;
@@ -326,9 +104,12 @@ export default async function SummaryPage({ params }: Props) {
     return <SignInRequired messages={translations.SignInRequired} />;
   }
 
+  // TODO: Make "from" a reserved word in the form generator
+  const { from: _, ...filterSearchParams } = searchParams;
+  const filters = buildDimensionFilters(filterSearchParams);
   const { data } = await getClient().query({
     query,
-    variables: { locale, eventSlug, surveySlug },
+    variables: { eventSlug, surveySlug, locale, filters },
   });
 
   if (!data.event?.forms?.survey?.summary) {
@@ -338,6 +119,7 @@ export default async function SummaryPage({ params }: Props) {
   const survey = data.event.forms.survey;
   const fields = survey.fields || [];
   const summary = survey.summary;
+  const dimensions = survey.dimensions || [];
 
   validateFields(fields);
   validateSummary(summary);
@@ -349,6 +131,8 @@ export default async function SummaryPage({ params }: Props) {
         {t.summaryTitle}
         <ViewHeading.Sub>{survey.title}</ViewHeading.Sub>
       </ViewHeading>
+      <DimensionFilters dimensions={dimensions} />
+      <p>{t.summaryOf(survey.countFilteredResponses, survey.countResponses)}</p>
       {fields.map((field) => (
         <SchemaFormField
           key={field.slug}
